MemoryCardFolder: Fix Wsign-compare warning.

+++ pcsx2/MemoryCardFolder.cpp
@@ -1008,7 +1008,7 @@ s32 FolderMemoryCard::Read(u8* dest, u32 adr, int size)
 			FolderMemoryCard::CalculateECC(ecc + (i * 3), &data[i * 0x80]);
 		}
 
-		pxAssert(size >= eccOffset);
+		pxAssert(static_cast<u32>(size) >= eccOffset);
 		const u32 copySize = std::min((u32)size - eccOffset, eccLength);
 		memcpy(dest + eccOffset, ecc, copySize);
 	}
